requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
confusable-homoglyphs==3.2.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-appconf==1.0.4
django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
django-classy-tags==2.0.0
django-crispy-forms==1.10.0
django-easy-maps==1.1.2
django-filter==2.4.0
django-geoposition==0.3.0
django-heroku==0.3.1
django-registration==3.1.1
django-username-email==2.4.1
django-utils-six==2.0
elasticsearch==7.10.1
filelock==3.0.12
Flask==1.1.2
geographiclib==1.50
geopy==2.1.0
geos==0.2.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
Pillow==8.1.0
pipenv==2020.11.15
postgis==1.0.4
psycopg2==2.8.6
python-dateutil==1.5
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
whitenoise==5.2.0
